requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
streamlit
pytesseract
pdf2img
sqlalchemy
pgvector
psycopg
mistralai
google-cloud-aiplatform
googlesearch-python
pycountry